limit the reference value for comparators

+++ PLC_esp8266/main/LogicProgram/Inputs/ComparatorGE.cpp
@@ -7,7 +7,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-ComparatorGE::ComparatorGE(int32_t reference, const MapIO io_adr, InputBase &prior_item)
+ComparatorGE::ComparatorGE(uint16_t reference, const MapIO io_adr, InputBase &prior_item)
     : ComparatorBase(reference, io_adr, prior_item) {
 }
